Refactoring of Theta Intersection.
Eliminated the ItersectionImplR class

Implemented the new standard way of handling nulls.

Changed the operation from "update" to "intersect"

We now have both stateful and stateless options.

Lots of code cleanup.
