+++ src/test/java/org/apache/datasketches/theta/CompactSketchTest.java
@@ -216,7 +216,7 @@ public class CompactSketchTest {
   @Test
   public void checkCompactCachePart() {
     //phony values except for curCount = 0.
-    long[] result = IntersectionImplR.compactCachePart(null, 4, 0, 0L, false);
+    long[] result = Intersection.compactCachePart(null, 4, 0, 0L, false);
     assertEquals(result.length, 0);
   }
